requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
click==8.2.1
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.3.2
openai==1.102.0
pgvector==0.4.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
